+++ .gitignore
@@ -5,7 +5,6 @@
 .idea/vcs.xml
 .idea/jsLibraryMappings.xml
 *.iml
-
 # Windows exe files
 *.exe
