+++ pom.xml
@@ -15,6 +15,7 @@
         <selenium.java.version>4.0.0</selenium.java.version>
         <webdrivermanager.version>4.4.3</webdrivermanager.version>
         <org.projectlombok.version>1.18.22</org.projectlombok.version>
+        <version>6.0.3</version>
     </properties>
 
     <dependencies>
@@ -55,14 +56,12 @@
         <dependency>
             <groupId>com.codeborne</groupId>
             <artifactId>selenide</artifactId>
-            <version>6.0.3</version>
             <scope>test</scope>
         </dependency>
 
         <dependency>
             <groupId>com.codeborne</groupId>
             <artifactId>selenide</artifactId>
-            <version>6.0.3</version>
             <scope>compile</scope>
         </dependency>
